Hackman Animationen in alle 4 Richtungen
