+++ src/Components/GameFieldComponent/FieldComponents/Corners/CornerLeftBottom.tsx
