+++ src/Components/GameFieldComponent/FieldComponents/Corners/CornerLeftTop.tsx
