+++ src/Components/GameFieldComponent/FieldComponents/HorizontalWalls/HorizontalWallLeftSideShort.tsx
