+++ src/Components/GameFieldComponent/FieldComponents/HorizontalWalls/HorizontalWallRightSideShort.tsx
