+++ src/Components/GameFieldComponent/FieldComponents/Path/Coin.tsx
