+++ src/Components/GameFieldComponent/FieldComponents/Path/Empty.tsx
