+++ src/Components/GameFieldComponent/FieldComponents/Path/Gate.tsx
