+++ src/Components/GameFieldComponent/FieldComponents/Path/Snack.tsx
