+++ src/Components/GameFieldComponent/FieldComponents/TPieces/TPieceLeft.tsx
