+++ src/Components/GameFieldComponent/FieldComponents/TPieces/TPieceRight.tsx
