+++ src/Components/GameFieldComponent/FieldComponents/VerticalWalls/VerticalWall.tsx
