+++ src/Components/GameFieldComponent/FieldComponents/VerticalWalls/VerticalWallBottomShort.tsx
