+++ src/Components/GameFieldComponent/FieldComponents/VerticalWalls/VerticalWallTopShort.tsx
