+++ src/Components/GameFieldComponent/GhostComponent/ghost.css
@@ -1,0 +1,39 @@
+.ghost{
+    z-index: 15;
+}
+.ghost-head {
+    position: absolute;
+    top: 10%;
+    left: 10%;
+    width: 70%;
+    height: 70%;
+    background: blue;
+    border-radius: 50% !important;
+    display: flex;
+    flex-direction: row;
+}
+.ghost-eye{
+    margin-left: 15%;
+    background-color: rgb(189, 170, 170);
+    height: 25%;
+    width: 20%;
+    border-radius: 50%;
+    transform: translateY(100%) translateX(30%);
+}
+.ghost-body {
+    position: absolute;
+    width: 70%;
+    height: 40%;
+    top: 45%;
+    left: 10%;
+    background-color: blue;
+    display: flex;
+    flex-direction: row;
+}
+.ghost-spikes{
+    width: 60%;
+    height: 80%;
+    border-radius: 50%;
+    background-color: black;
+    transform: translateY(100%);
+}
