+++ src/Components/GameFieldComponent/gameField.css
@@ -1,0 +1,13 @@
+.row {
+    position: relative;
+    margin: 0 auto;
+    width: fit-content;
+}
+
+.field {
+    position: relative;
+    height: 100%;
+    background-color: black;
+    color: black;
+    display: inline-block;
+}
