--- src/Components/Hackman/hackman.css
@@ -1,95 +1,0 @@
-/* Hackman */
-.hackmanForm {
-    position: relative;
-    width: 80%;
-    height: 80%;
-    border-radius: 50%;
-    z-index: 99;
-}
- .hackmanFill{
-     background:conic-gradient(#ff9000 0deg 360deg);
-
- }   
-.upView{
-    transform: rotateZ(90deg) translateX(10%) translateY(-12%);
-}
-.bottomView{
-    transform: rotateZ(-90deg) translateX(-10%) translateY(12%);
-}
-.leftView{
-    transform: rotateY(0deg) translateX(10%) translateY(12%);
-}
-.rightView{
-    transform: rotateY(-180deg) translateX(-10%) translateY(12%);
-}
-.move-from-bottom-to-up{
-    animation-name: move-from-bottom-to-top;
-    animation-duration: 0.25s;   
-    -webkit-animation-name: move-from-bottom-to-top; /* Chrome, Safari, Opera */
-    -webkit-animation-duration: 0.25s; /* Chrome, Safari, Opera */
-}
-.move-from-up-to-bottom{
-    animation-name: move-from-up-to-bottom;
-    animation-duration: 0.25s;   
-    -webkit-animation-name: move-from-up-to-bottom; /* Chrome, Safari, Opera */
-    -webkit-animation-duration: 0.25s; /* Chrome, Safari, Opera */ 
-}
-.move-from-left-to-right{
-    animation-name: move-from-left-to-right;
-    animation-duration: 0.25s;   
-    -webkit-animation-name: move-from-left-to-right; /* Chrome, Safari, Opera */
-    -webkit-animation-duration: 0.25s; /* Chrome, Safari, Opera */
-}
-.move-from-right-to-left{
-    animation-name: move-from-right-to-left;
-    animation-duration: 0.25s;   
-    -webkit-animation-name: move-from-right-to-left; /* Chrome, Safari, Opera */
-    -webkit-animation-duration: 0.25s; /* Chrome, Safari, Opera */
-}
-.eye{
-    background-color: #19181A;
-    width: 18%;
-    height: 18%;
-    border-radius: 50%;
-    transform: translateX(120%) translateY(100%);
-    z-index: 7;
-}
-.eye-move{
-    transform: translateX(110%) translateY(60%) ;
-}
-
-
-@keyframes move-from-right-to-left {
-    0%  {background:conic-gradient(#ff9000 0deg 360deg);transform: translateX(100%) translateY(12%);}
-    7.15%   {background:conic-gradient(#ff9000 265deg, #000 265deg 275deg,#ff9000 275deg 360deg);transform: translateX(75%) translateY(12%);}
-    14.3%   {background:conic-gradient(#ff9000 260deg, #000 260deg 280deg,#ff9000 280deg 360deg);transform: translateX(70%) translateY(12%);}
-    21.45%   {background:conic-gradient(#ff9000 255deg, #000 255deg 285deg,#ff9000 285deg 360deg);transform: translateX(60%) translateY(12%);}
-    28.6%   {background:conic-gradient(#ff9000 250deg, #000 250deg 290deg,#ff9000 290deg 360deg);transform: translateX(55%) translateY(12%);}
-    35.75%   {background:conic-gradient(#ff9000 245deg, #000 245deg 295deg,#ff9000 295deg 360deg);transform: translateX(45%) translateY(12%);}
-    42.8%   {background:conic-gradient(#ff9000 240deg, #000 240deg 300deg,#ff9000 300deg 360deg);transform: translateX(40%) translateY(12%);}
-    50%  {background:conic-gradient(#ff9000 235deg, #000 235deg 305deg,#ff9000 305deg 360deg);transform: translateX(30%) translateY(12%);}
-    57.15%   {background:conic-gradient(#ff9000 240deg, #000 240deg 300deg,#ff9000 300deg 360deg);transform: translateX(25%) translateY(12%);}
-    64.3%   {background:conic-gradient(#ff9000 245deg, #000 245deg 295deg,#ff9000 295deg 360deg);transform: translateX(20%) translateY(12%);}
-    71.45%   {background:conic-gradient(#ff9000 250deg, #000 250deg 290deg,#ff9000 290deg 360deg);transform: translateX(15%) translateY(12%);}
-    78.6%   {background:conic-gradient(#ff9000 255deg, #000 255deg 285deg,#ff9000 285deg 360deg);transform: translateX(10%) translateY(12%);}
-    85.75%   {background:conic-gradient(#ff9000 260deg, #000 260deg 280deg,#ff9000 280deg 360deg);transform: translateX(5%) translateY(12%);}
-    92.9%   {background:conic-gradient(#ff9000 265deg, #000 265deg 275deg,#ff9000 275deg 360deg);transform: translateX(0%) translateY(12%);}
-    100%  {background:conic-gradient(#ff9000 0deg 360deg);transform: translateX(0%) translateY(12%);} 
-  } 
-@-webkit-keyframes move-from-right-to-left {
-    0%  {background:conic-gradient(#ff9000 0deg 360deg);transform: translateX(100%) translateY(12%);}
-    7.15%   {background:conic-gradient(#ff9000 265deg, #000 265deg 275deg,#ff9000 275deg 360deg);transform: translateX(75%) translateY(12%);}
-    14.3%   {background:conic-gradient(#ff9000 260deg, #000 260deg 280deg,#ff9000 280deg 360deg);transform: translateX(70%) translateY(12%);}
-    21.45%   {background:conic-gradient(#ff9000 255deg, #000 255deg 285deg,#ff9000 285deg 360deg);transform: translateX(60%) translateY(12%);}
-    28.6%   {background:conic-gradient(#ff9000 250deg, #000 250deg 290deg,#ff9000 290deg 360deg);transform: translateX(55%) translateY(12%);}
-    35.75%   {background:conic-gradient(#ff9000 245deg, #000 245deg 295deg,#ff9000 295deg 360deg);transform: translateX(45%) translateY(12%);}
-    42.8%   {background:conic-gradient(#ff9000 240deg, #000 240deg 300deg,#ff9000 300deg 360deg);transform: translateX(40%) translateY(12%);}
-    50%  {background:conic-gradient(#ff9000 235deg, #000 235deg 305deg,#ff9000 305deg 360deg);transform: translateX(30%) translateY(12%);}
-    57.15%   {background:conic-gradient(#ff9000 240deg, #000 240deg 300deg,#ff9000 300deg 360deg);transform: translateX(25%) translateY(12%);}
-    64.3%   {background:conic-gradient(#ff9000 245deg, #000 245deg 295deg,#ff9000 295deg 360deg);transform: translateX(20%) translateY(12%);}
-    71.45%   {background:conic-gradient(#ff9000 250deg, #000 250deg 290deg,#ff9000 290deg 360deg);transform: translateX(15%) translateY(12%);}
-    78.6%   {background:conic-gradient(#ff9000 255deg, #000 255deg 285deg,#ff9000 285deg 360deg);transform: translateX(10%) translateY(12%);}
-    85.75%   {background:conic-gradient(#ff9000 260deg, #000 260deg 280deg,#ff9000 280deg 360deg);transform: translateX(5%) translateY(12%);}
-    92.9%   {background:conic-gradient(#ff9000 265deg, #000 265deg 275deg,#ff9000 275deg 360deg);transform: translateX(0%) translateY(12%);}
-    100%  {background:conic-gradient(#ff9000 0deg 360deg);transform: translateX(0%) translateY(12%);} 
-  }
