+++ src/Components/StatsComponent/Stats.tsx
@@ -1,7 +1,5 @@
 import { useSelector } from "react-redux";
 import { State } from "../../State/store";
-import './stats.css';
-
 const Stats: React.FC<any> = () => {
     const eatenCoins = useSelector((state:State)=>state.eatenCoins)
     const remainingLives:number = 3;
