--- src/Interfaces/CharacterProps.ts
@@ -1,5 +1,0 @@
-import Direction from "../Types/Direction";
-
-export type CharacterProps = {
-    richtung: Direction;
-  };
