+++ src/SpielfeldLayout.tsx
@@ -1,22 +1,22 @@
-import Coin from "./Components/Fields/Path/Coin";
-import Empy from "./Components/Fields/Path/Empty";
-import Holl from "./Components/Fields/HorizontalWalls/HorizontalWall";
-import Hwrs from "./Components/Fields/HorizontalWalls/HorizontalWallRightSideShort";
-import Hwls from "./Components/Fields/HorizontalWalls/HorizontalWallLeftSideShort";
-import Vell from "./Components/Fields/VerticalWalls/VerticalWall";
-import Vwts from "./Components/Fields/VerticalWalls/VerticalWallTopShort";
-import Vwbs from "./Components/Fields/VerticalWalls/VerticalWallBottomShort";
-import Tpbo from "./Components/Fields/TPieces/TPieceBottom";
-import Tpto from "./Components/Fields/TPieces/TPieceTop";
-import Tprs from "./Components/Fields/TPieces/TPieceRight";
-import Tpls from "./Components/Fields/TPieces/TPieceLeft";
-import Colt from "./Components/Fields/Corners/CornerLeftTop";
-import Colb from "./Components/Fields/Corners/CornerLeftBottom";
-import Cort from "./Components/Fields/Corners/CornerRightTop";
-import Corb from "./Components/Fields/Corners/CornerRightBottom";
-import Snac from "./Components/Fields/Path/Snack";
-import Hack from "./Components/Hackman/Hackman";
-import Geis from "./Components/Ghost/Ghost";
+import Coin from "./Components/GameFieldComponent/FieldComponents/Path/Coin";
+import Empy from "./Components/GameFieldComponent/FieldComponents/Path/Empty";
+import Holl from "./Components/GameFieldComponent/FieldComponents/HorizontalWalls/HorizontalWall";
+import Hwrs from "./Components/GameFieldComponent/FieldComponents/HorizontalWalls/HorizontalWallRightSideShort";
+import Hwls from "./Components/GameFieldComponent/FieldComponents/HorizontalWalls/HorizontalWallLeftSideShort";
+import Vell from "./Components/GameFieldComponent/FieldComponents/VerticalWalls/VerticalWall";
+import Vwts from "./Components/GameFieldComponent/FieldComponents/VerticalWalls/VerticalWallTopShort";
+import Vwbs from "./Components/GameFieldComponent/FieldComponents/VerticalWalls/VerticalWallBottomShort";
+import Tpbo from "./Components/GameFieldComponent/FieldComponents/TPieces/TPieceBottom";
+import Tpto from "./Components/GameFieldComponent/FieldComponents/TPieces/TPieceTop";
+import Tprs from "./Components/GameFieldComponent/FieldComponents/TPieces/TPieceRight";
+import Tpls from "./Components/GameFieldComponent/FieldComponents/TPieces/TPieceLeft";
+import Colt from "./Components/GameFieldComponent/FieldComponents/Corners/CornerLeftTop";
+import Colb from "./Components/GameFieldComponent/FieldComponents/Corners/CornerLeftBottom";
+import Cort from "./Components/GameFieldComponent/FieldComponents/Corners/CornerRightTop";
+import Corb from "./Components/GameFieldComponent/FieldComponents/Corners/CornerRightBottom";
+import Snac from "./Components/GameFieldComponent/FieldComponents/Path/Snack";
+import Hack from "./Components/GameFieldComponent/HackmanComponent/Hackman";
+import Geis from "./Components/GameFieldComponent/GhostComponent/Ghost";
 import React from 'react';
 
 const SpielfeldLayout = () => {
