+++ src/Tests/UtilityFunctions/GameFieldService/MoveHackman/MoveHackman.test.ts
@@ -1,5 +1,5 @@
-import Empty from '../../../../Components/Fields/Path/Empty';
-import Hackman from '../../../../Components/Hackman/Hackman';
+import Empty from '../../../../Components/GameFieldComponent/FieldComponents/Path/Empty';
+import Hackman from '../../../../Components/GameFieldComponent/HackmanComponent/Hackman';
 import { hackmanMovesDown, hackmanMovesLeft, hackmanMovesRight, hackmanMovesUp } from '../../../../UtilityFunctions/move/MoveHackman';
 import GetHackmanByPosition from '../../../Mock/Character.Mock';
 import { testGameFieldCoinHackmanMiddle } from '../../../Mock/TestField.Mock';
