+++ src/UtilityFunctions/move/CanMove.ts
@@ -1,7 +1,7 @@
-import Coin from "../../Components/Fields/Path/Coin";
-import Empty from "../../Components/Fields/Path/Empty";
-import Ghost from "../../Components/Ghost/Ghost";
-import Snack from "../../Components/Fields/Path/Snack";
+import Coin from "../../Components/GameFieldComponent/FieldComponents/Path/Coin";
+import Empty from "../../Components/GameFieldComponent/FieldComponents/Path/Empty";
+import Ghost from "../../Components/GameFieldComponent/GhostComponent/Ghost";
+import Snack from "../../Components/GameFieldComponent/FieldComponents/Path/Snack";
 import Moveable from "../../Types/Moveable";
 import Coordinate from "../../Types/Coordinate";
 import Direction from "../../Types/Direction";
